requirements.txt
pytest==8.0.2
Appium-Python-Client==3.2.1
